Added LivePreview class
The LivePreview class itself is responsible for the serial input and
thus takes a port and a baudrate.  Further it requires a feature
extractor, that can be fed with the input coming from the serial
interface until it provides a complete set of features, and a model
that takes the features provided by the feature extractor and returns
its prediction.  The feature extractor and model have to be compatible.