Fix hanging presses
Sometimes when you release to key presses at the same time they are
apparently not both handled by the `touchEnded` function which leads
to presses which are never released.  To avoid this we check in the
main loop if there are any presses left and if there are not we just
reset the map.