Changed build partition to min_spiffs.csv
Adding the BLESerial library made the program to big so pio threw an error. This could be fixed by changing the partition config to min_spiffs.csv, which reduces the size allocated for a file system and therefore increased available program-size, as the ESP32 probably just has 4mb of flash memory.