Avoid unusable stand-alone data
In some datasets there the virtual keyboard has some data recorded
before the board could collect the respective data from the sensors.

This is generally not a big problem as this data is discarded when the
dataset is preprocessed but it could theoretically cause some strange
data in the beginning and clutter the joined ascii output.

Additionally, this commit includes a newer version of the
`requirements.txt`.