Static Dataset Split
Instead of splitting the dataset into train, dev and test distribution
when building the model, we perform a static split when preparing the
dataset.

This is supposed to make our research results more consistent and
reproducible.

Although unbalanced (this might cause an issue), we have a farely
large dataset; thus, we chose the following split:
- 70% train
- 18% dev
- 12% test