Actually switching the order of the sensors
The Sensors array and everything keeps it's order, but the sensor_presence array now stores the channel on which the sensor is and a -1 if it's missing.

The function tca_select is further abstracted in this file now, by a new function select_sensor which first gets the channel from the array by id.

The presence checks were replace by a comparison with -1.

Another function was added: A setup_senors function which is called in main.cpp which fills the sensor_presence array with the appropriate channel depending on the RIGHT_HAND variable and then sets up the sensors in the standard order.