Change internal representation of key position
The former method of representing a key position was a string with the
character representation of the column and row position.  This is
neither extensible nor is it a very clean way to represent the
position.  Thus, we changed the representation to a 2-tuple (column,
row).

Additionally, we changed the column and row values to match those that
qmk uses internally.