Check for global hid_listen and make local script location relative
Calling the `keylogger-for-data-collection.sh` script from another
directory will result in an error because the local hid_listen script
will not be in that directory.  Referring to the hid_listen script
from the keylogger script location (dirname "$0") solves this issue.

Further checking for a global hid_listen script makes things easier
with a nix-shell environment.