First draft of all rings
Created a ring module that takes the finger diameter and optionally a
tolerence and an angle at which the center of the hole is shifted to.

With this five rings for each individual finger were added easily. What
is missing are the fillets and correct parameters.

Another thing that could be improved is the way the SenseUnit slot is
attached to the ring. When the diameter of the outer contour of the ring
is smaller than the width of SenseUnit and the walls there are edges on
the ring.