Fix buggy pen thickness
When changing the size of the canvas buffer, sometimes the pen
thickness would arbitrarily double. It turns out that the `on-paint`
callback is called when another such callback has yet to finish.

To fix this we associate a default pen with the component on the first
`on-paint` callback and continue using that pen instead of retaining
the current pen in every callback.