Use python class system for data preparation abstraction
Before we imported functions based on file names which is not very
reliable and a bit hacky. So instead we now abstract data preparation
to instances of the `DataPreparation` class.

NOTE: Externally the software should behave the same way it did
before. Though, I should note that the data type can now also be
specified via stdin if the respective environment variable is not set.