Bug live-plotting: Changing plot min/max
When changing min/max values of the plot, the other one(s) got reset to
the default value. This was fixed by setting the int-[min/max/size]
variable to the new value upon update. Before it stayed at the default
value and got used if one of the other fields got updated.