machine-learning: Add window-based naive prediction strategy
This commit mainly adds a third strategy for the naive prediction
idea which uses a rolling window to compute the maximal difference
between maximum and minimum of the window.

This commit also improves upon the loading code block which now also
supports choosing the stage to load if the user uses a prefix
argument.